+++ src/internal.cpp
@@ -14,10 +14,13 @@ Internal::Internal ()
       external_prop_is_lazy (true), rephased (0), vsize (0), max_var (0),
       clause_id (0), original_id (0), reserved_ids (0), conflict_id (0),
       level (0), vals (0), score_inc (1.0), scores (this), conflict (0),
-      ignore (0), external_reason (&external_reason_clause), notified (0),
+      ignore (0), external_reason (&external_reason_clause),
+      newest_clause (0), force_no_backtrack (false), from_propagator (false),
+      tainted_literal (0), notified (0), probe_reason (0),
       propagated (0), propagated2 (0), propergated (0), best_assigned (0),
       target_assigned (0), no_conflict_until (0), unsat_constraint (false),
-      marked_failed (true), proof (0), checker (0), tracer (0),
+      marked_failed (true), multitrail_dirty (0), num_assigned (0),
+      proof (0), checker (0), tracer (0),
       lratchecker (0), lratbuilder (0), opts (this),
 #ifndef QUIET
       profiles (this), force_phase_messages (false),
@@ -45,6 +48,7 @@ Internal::~Internal () {
     vals -= vsize;
     delete[] vals;
   }
+  clear_trails (0);
 }
 
 /*------------------------------------------------------------------------*/
@@ -173,10 +177,7 @@ void Internal::add_original_lit (int lit) {
       // Use the external form of the clause for printing in proof
       // Externalize(internalized literal) != external literal
       assert (!original.size () || !external->eclause.empty ());
-      if (opts.lrat)
-        proof->add_external_original_clause (id, external->eclause);
-      else
-        proof->add_external_original_clause (id, external->eclause);
+      proof->add_external_original_clause (id, external->eclause);
     }
     add_new_original_clause (id);
     original.clear ();
@@ -665,6 +666,10 @@ int Internal::local_search () {
 
 /*------------------------------------------------------------------------*/
 
+// if preprocess_only is false and opts.ilb is true we do not preprocess
+// such that we do not have to backtrack to level 0.
+// TODO: check restore_clauses works on higher level
+//
 int Internal::solve (bool preprocess_only) {
   assert (clause.empty ());
   START (solve);
@@ -674,6 +679,8 @@ int Internal::solve (bool preprocess_only) {
     LOG ("internal solving in full mode");
   init_report_limits ();
   int res = already_solved ();
+  if (!res && preprocess_only && level)
+    backtrack ();
   if (!res)
     res = restore_clauses ();
   if (!res) {
@@ -681,12 +688,12 @@ int Internal::solve (bool preprocess_only) {
     if (!preprocess_only)
       init_search_limits ();
   }
-  if (!res)
+  if (!res && !level)
     res = preprocess ();
   if (!preprocess_only) {
-    if (!res)
+    if (!res && !level)
       res = local_search ();
-    if (!res)
+    if (!res && !level)
       res = lucky_phases ();
     if (!res || (res == 10 && external_prop)) {
       if (res == 10 && external_prop && level)
@@ -707,9 +714,9 @@ int Internal::already_solved () {
     LOG ("already inconsistent");
     res = 20;
   } else {
-    if (level)
+    if (level && !opts.ilb)
       backtrack ();
-    if (!propagate ()) {
+    if (!opts.ilb && !propagate ()) {
       LOG ("root level propagation produces conflict");
       learn_empty_clause ();
       res = 20;