+++ src/probe.cpp
@@ -299,6 +299,8 @@ inline void Internal::probe_assign (int lit, int parent) {
   Var &v = var (idx);
   v.level = level;
   v.trail = (int) trail.size ();
+  assert ((int) num_assigned < max_var);
+  num_assigned++;
   v.reason = level ? probe_reason : 0;
   probe_reason = 0;
   set_parent_reason_literal (lit, parent);
@@ -312,6 +314,8 @@ inline void Internal::probe_assign (int lit, int parent) {
   assert (val (lit) > 0);
   assert (val (-lit) < 0);
   trail.push_back (lit);
+  if (!level && opts.reimply && external_prop && !external_prop_is_lazy)
+    notify_trail.push_back (lit);
 
   // Do not save the current phase during inprocessing but remember the
   // number of units on the trail of the last time this literal was