+++ src/vivify.cpp
@@ -54,6 +54,8 @@ inline void Internal::vivify_assign (int lit, Clause *reason) {
   Var &v = var (idx);
   v.level = level;               // required to reuse decisions
   v.trail = (int) trail.size (); // used in 'vivify_better_watch'
+  assert ((int) num_assigned < max_var);
+  num_assigned++;
   v.reason = level ? reason : 0; // for conflict analysis
   if (!level)
     learn_unit_clause (lit);
@@ -158,7 +160,7 @@ bool Internal::vivify_propagate () {
             j--;
           } else if (!u) {
             assert (v < 0);
-            build_chain_for_units (other, w.clause, 0);
+            vivify_chain_for_units (other, w.clause);
             vivify_assign (other, w.clause);
             lrat_chain.clear ();
           } else {
@@ -1037,6 +1039,29 @@ void Internal::vivify_build_lrat (int lit, Clause *reason) {
     if (v.reason) { // recursive justification
       vivify_build_lrat (other, v.reason);
     }
+  }
+  lrat_chain.push_back (reason->id);
+}
+
+// calculate lrat_chain
+//
+inline void Internal::vivify_chain_for_units (int lit, Clause *reason) {
+  if (!opts.lrat || opts.lratexternal)
+    return;
+  // LOG ("building chain for units");        bad line for debugging
+  // equivalence if (opts.chrono && assignment_level (lit, reason)) return;
+  if (level)
+    return; // not decision level 0
+  assert (lrat_chain.empty ());
+  for (auto &reason_lit : *reason) {
+    if (lit == reason_lit)
+      continue;
+    assert (val (reason_lit));
+    if (!val (reason_lit))
+      continue;
+    const unsigned uidx = vlit (val (reason_lit) * reason_lit);
+    uint64_t id = unit_clauses[uidx];
+    lrat_chain.push_back (id);
   }
   lrat_chain.push_back (reason->id);
 }